refactor: tweak translation keys, etc (#45)
