+++ packages/ui/public/locales/en/translation.json
@@ -89,7 +89,7 @@
   "No accounts selected": "",
   "No dapps meet search query:": "",
   "Only connect if you trust the application": "",
-  "Open Coong Wallet on another device and scan this QR Code to transfer your {{exportObject}}.": "",
+  "Open Coong Wallet on another device and scan this QR Code to transfer your {{object}}.": "",
   "Open account controls": "",
   "Open settings": "",
   "Password changed successfully!": "",
@@ -144,7 +144,7 @@
   "Unlock your wallet": "",
   "View Secret Recovery Phrase": "",
   "Wallet Access Request": "",
-  "{{exportObject}} Export QR Code": "",
+  "{{object}} Export QR Code": "",
   "Wallet imported": "",
   "Wallet is already initialized": "",
   "Wallet password": "",
