+++ packages/ui/src/components/pages/Accounts/AccountControls/ExportAccountDialog.tsx
@@ -6,8 +6,8 @@ import { AccountQrBackup } from '@coong/keyring/types';
 import { TabContext, TabList, TabPanel } from '@mui/lab';
 import { Dialog, DialogContent, Tab } from '@mui/material';
 import DialogTitle from 'components/shared/DialogTitle';
-import Json from 'components/shared/Json';
-import QrCode from 'components/shared/QrCode';
+import Json from 'components/shared/export/Json';
+import QrCode from 'components/shared/export/QrCode';
 import VerifyingPasswordForm from 'components/shared/forms/VerifyingPasswordForm';
 import useDialog from 'hooks/useDialog';
 import { useWalletState } from 'providers/WalletStateProvider';
